+++ packages/remix-express/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@remix-run/express`
 
+## 1.19.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/node@1.19.3`
+
 ## 1.19.2
 
 ### Patch Changes